moved register_database_notifciation_request to correct place. as referred to in:
https://github.com/SpiNNakerManchester/sPyNNaker8/issues/37
https://github.com/SpiNNakerManchester/sPyNNaker7/issues/20
